// src/app/products/models/products/product.model.ts
export interface Product {
    ProductId?: string;
    Category: string;
    MainCategory: string;
    TaxTarifCode :number;
    SupplierName: string;
    WeightMeasure: number;
    WeightUnit:string;
    Description: string;
    Name: string;
    DateOfSale:string;
    ProductPicUrl: string;
    Status: string;
    Quantity : number;
    UoM: number;
    CurrencyCode: string;
    Price: number;
    Width: number;
    Depth: number;
    Height : number;
    DimUnit: string

}